将 normal() 函数的输入参数由 int 改为 long long，解决数据溢出的 Bug。

+++ src/array/string/HashedString.hpp
@@ -48,7 +48,7 @@ namespace laozhou {
             return val;
         }
     private:
-        long long normal(int x){
+        long long normal(long long x){
             assert(_mo>0);
             return (x%_mo+_mo)%_mo;
         }
